# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Operator360 Portal Backend API
  version: 1.0.0
  description: |
    Contract reference for the Operator360 portal backend, generated from the
    current handler implementations on branch f/oid-pid. All routes are under
    `/api` and require a bearer/session token validated by AuthMiddleware.

    Endpoints whose S3 file paths are keyed by an operator now resolve the S3
    prefix via `opt_master.data_path` looked up by `opt_id` — they no longer
    accept or require `opt_state` / `opt_district`.
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required: [error]
    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_records:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_previous:
          type: boolean
...
  responses:
    Unauthorized:
      description: User not found in context
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    ServerError:
      description: Internal server error (S3, DB, or parse failure)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  parameters:
    OptId:
      name: opt_id
      in: query
      required: true
      schema:
        type: string
      example: MH_WMIT_SN_NS046496
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
...
paths:
...
  /search_operator_packets:
    get:
      summary: Search/paginate an operator's SID packets (sid.parquet)
      description: |
        File path is resolved via opt_master.data_path looked up by opt_id.
        opt_state/opt_district are no longer accepted.
      operationId: SearchOperatorPacketsBySID
      parameters:
        - $ref: "#/components/parameters/OptId"
        - $ref: "#/components/parameters/Page"
        - name: page_size
          in: query
          schema: { type: integer, default: 10, maximum: 1000 }
        - name: sid
          in: query
          schema: { type: string }
          description: Substring match against the eid field.
        - name: anomaly_filter
          in: query
          schema: { type: string, enum: [anomalous, non-anomalous] }
        - name: enrollment_type
          in: query
          schema: { type: string, enum: [new_enrollment, update] }
        - name: date
          in: query
          schema: { type: string }
          example: "2025-12-25"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  regional_office: { type: string }
                  operator_id: { type: string }
                  file_path: { type: string }
                  search_sid: { type: string }
                  anomaly_filter: { type: string }
                  enrollment_type_filter: { type: string }
                  date_filter: { type: string }
                  pagination: { $ref: "#/components/schemas/Pagination" }
                  count: { type: integer }
                  data:
                    type: array
                    items: { type: object, additionalProperties: true }
                  requested_by: { type: string }
        "400":
          description: opt_id missing
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Operator data path or sid.parquet not found
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "500": { $ref: "#/components/responses/ServerError" }
...